Extract ReaderPostUiStateBuilder from ReaderDiscoverViewModel
